requirements.txt
azure-cognitiveservices-vision-customvision
msrest
python-dotenv